pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "gensim>=4.3.3",
    "google-generativeai>=0.8.4",
    "matplotlib>=3.10.1",
    "nltk>=3.9.1",
    "numpy>=1.26.4",
    "pandas>=2.2.3",
    "plotly>=6.0.1",
    "streamlit>=1.43.2",
    "wordcloud>=1.9.4",
]
